requirements.txt
######################################
# 🚀 CORE ENVIRONMENT & UTILITIES
######################################
setuptools==70.0.0
typing_extensions==4.14.0
typing-inspection==0.4.1
platformdirs==4.3.8
packaging==25.0
colorama==0.4.6
lazy_loader==0.4
pluggy==1.6.0
shellingham==1.5.4
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
requests==2.32.4
tqdm==4.67.1
exceptiongroup==1.3.0

######################################
# 🤪 TESTING & DEBUGGING
######################################
pytest==8.4.0
iniconfig==2.1.0

######################################
# 🤠 MACHINE LEARNING & DSP
######################################
numpy==2.2.6
scipy==1.15.3
scikit-learn==1.7.0
librosa==0.11.0
soundfile==0.13.1
numba==0.61.2
llvmlite==0.44.0
joblib==1.5.1
soxr==0.5.0.post1
pooch==1.8.2
audioread==3.0.1

######################################
# 🎮 VIDEO & AUDIO PROCESSING
######################################
moviepy==2.2.1
ffmpeg-python==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
proglog==0.1.12
audioop-lts==0.2.1
pytesseract==0.3.13
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86

######################################
# 🌐 FLASK SERVER + API
######################################
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
python-dotenv==1.1.0

######################################
# 🌍 FASTAPI / STARLETTE STACK
######################################
starlette==0.47.0
uvicorn==0.34.3
sse-starlette==2.3.6
httpx==0.28.1
httpcore==1.0.9
sniffio==1.3.1
httpx-sse==0.4.0
anyio==4.9.0

######################################
# 🔐 AUTHENTICATION
######################################
Authlib==1.6.0
cryptography==45.0.4
cffi==1.17.1
pycparser==2.22
blinker==1.9.0

######################################
# 🧹 YOUTUBE / CONTENT DOWNLOADING
######################################
yt-dlp==2025.6.9

######################################
# 📊 RENDERING / UI / TEXT STYLING
######################################
rich==14.0.0
Pygments==2.19.1
markdown-it-py==3.0.0
mdurl==0.1.2

######################################
# 🧼 Pydantic & Config
######################################
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
openapi-pydantic==0.5.1

######################################
# ⚙️ OTHER DEPENDENCIES
######################################
mcp==1.9.4
fastmcp==2.8.0
jiter==0.10.0
python-multipart==0.0.20
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
msgpack==1.1.1
distro==1.9.0